@@ -6,22 +6,21 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  email_validator: ^1.0.6
-  intl: ^0.16.1
-  flutter_launcher_icons: ^0.7.0
-  search_widget: ^1.0.2
+  email_validator: ^2.0.1
+  flutter_launcher_icons: ^0.9.0
   url_launcher: ^6.0.3
   fluttertoast: ^8.0.6
   flutter_swiper_null_safety: ^1.0.1
   font_awesome_flutter: ^9.0.0
-  flutter_login: ^1.1.0
+  flutter_login: ^2.0.0-nullsafety.0
+  grpc: ^3.0.0
 
 dev_dependencies:
   flutter_test:
